add generic nested results visitor, rewrite xxh visitor
